# Fix for memory issue when watermarking images