I rewrote the image watermarking because of the related memory issue I had. See also #138. 

Few remarks:
- not sure what the specs are for watermarking, I took another approach to watermark an image with another image. Could be that it is not according the specs.
- ~~when running the test (on OSX) I get an error regarding defering a C-pointer:~~

- I run some benchmarks already available see the differences below.
Benchmarks (ran on OSX, local machine):
```
// New implementation
BenchmarkWatermarkJpeg-4        	      20	  60215874 ns/op
BenchmarkWatermarPng-4          	     500	   3406345 ns/op
BenchmarkWatermarWebp-4         	      30	  40243090 ns/op
BenchmarkWatermarkImageJpeg-4   	      20	  72954939 ns/op
BenchmarkWatermarImagePng-4     	      30	  38325610 ns/op
BenchmarkWatermarImageWebp-4    	      20	  70658176 ns/op

// Previous implementation
BenchmarkWatermarkJpeg-4        	      20	  60050577 ns/op
BenchmarkWatermarPng-4          	     500	   3385920 ns/op
BenchmarkWatermarWebp-4         	      30	  39207303 ns/op
BenchmarkWatermarkImageJpeg-4   	      10	 148488564 ns/op
BenchmarkWatermarImagePng-4     	      30	  47970915 ns/op
BenchmarkWatermarImageWebp-4    	      20	  66507471 ns/op
```

- good news is that the memory issues are solved, and I can use this library to watermark images 👍