@jaume-pinyol 2017-03-31
Sorry for not answering, I didn't see it. Thanks for fixing the leak and improving the code ;) LGTM 👍