@ajdevries 2017-03-31
Also fixed the `unref` issue I was having.