@h2non 2017-04-05
This looks good to me. Just out of curiosity: what `libvips` version are you running in local?