@ajdevries 2017-04-05
That is a good one, only tested it with `libvips 8.4.5`. No earlier versions. I think it works with earlier versions, but not sure.