@h2non 2017-04-05
I was too much optimistic here. `vips_bandjoin_const1` function is not available in previous `libvips` versions. I'm going to revert this. We should provide better cross-version compability. In case that watermark code is not API supported in the runtime `libvips` version, we should return an error at runtime, but not fail at compilation time.