@h2non 2017-04-05
Feel free to create a new PR once the improvements are done! Thanks a lot.