@ajdevries 2017-04-05
Hmmm. But that isn't new, and the method is also there in the old watermark code (in master):
- https://github.com/h2non/bimg/blob/master/vips.h#L461